Unit test amendments
* getEnvIndex functionality now included in getEnvironmnts
* Moved getEnvironments to helpers
* related unit tests updated
