+++ test/index.test.js
@@ -1,36 +1,4 @@
-import Deviance, { getEnvIndex, getEnvironment } from '../src/index';
-
-describe('Given getEnvIndex', () => {
-    describe('When no envFlag passed', () => {
-        const args = [];
-        it('Then returns result of 0', () => {
-            expect(getEnvIndex(args)).toBe(0);
-        });
-    });
-
-    describe('When envFlag passed as first set of arguments', () => {
-        const args = ['-mock', 'mocked', '-e', 'value'];
-        it('Then returns index of "value"', () => {
-            expect(getEnvIndex(args)).toBe(3);
-        });
-    });
-});
-
-describe('Given getEnvironment', () => {
-    describe('When no envFlag passed', () => {
-        const args = [];
-        it('Then returns value of default', () => {
-            expect(getEnvironment(args)).toBe('default');
-        });
-    });
-
-    describe('When envFlag passed', () => {
-        const args = ['-e', 'value'];
-        it('Then returns value of passed string', () => {
-            expect(getEnvironment(args)).toBe('value');
-        });
-    });
-});
+import Deviance from '../src/index';
 
 describe('Given Deviance class', () => {
     describe('When not passing in any custom settings', () => {
@@ -40,7 +8,7 @@ describe('Given Deviance class', () => {
         });
         it('Then default reporter enabled setting is false', () => {
             expect(instance.settings.reporting.enabled).toBe(false);
-        })
+        });
     });
 
     describe('When passing custom path settings', () => {
@@ -51,7 +19,8 @@ describe('Given Deviance class', () => {
         };
         const instance = new Deviance(settings);
         it('Then custom settings overwrite default settings', () => {
-            expect(instance.settings.regression.expectedPath).toContain(settings.regression.expectedPath);
+            expect(instance.settings.regression.expectedPath)
+                .toContain(settings.regression.expectedPath);
         });
     });
 
@@ -75,13 +44,16 @@ describe('Given Deviance class', () => {
     });
 
     describe('When env const equals a value', () => {
-        const instance = new Deviance({}, ['-e', 'mockValue']);
+        process.argv.push('-e', 'mockValue');
+        const instance = new Deviance();
         it('Then regression expectedPath is contains envFlag', () => {
-            expect(instance.settings.regression.expectedPath).toContain('mockValue');
+            expect(instance.settings.regression.expectedPath)
+                .toContain('mockValue');
         });
 
         it('Then regression actualPath is contains envFlag', () => {
-            expect(instance.settings.regression.actualPath).toContain('mockValue');
+            expect(instance.settings.regression.actualPath)
+                .toContain('mockValue');
         });
     });
 });
